Reorganized modifiers location, better code organization.
