+++ src/main/java/tcintegrations/client/compat/Create.java
@@ -6,7 +6,7 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
 
-import tcintegrations.items.armor.modifiers.EngineersGogglesModifier;
+import tcintegrations.items.modifiers.armor.EngineersGogglesModifier;
 
 public class Create {
