+++ src/main/java/tcintegrations/items/modifiers/tool/ElementalModifier.java
@@ -1,4 +1,4 @@
-package tcintegrations.items.tool.modifiers;
+package tcintegrations.items.modifiers.tool;
 
 import java.util.List;
